fix: Diff listing file highlighting slightly off during scroll
The Intersection Observer margins aren't exact since they're partially based off of the viewport height. Previously the point where the hightlight would switch files was set to the top of the page `100%`, which wasn't always correct if the PR sticky nav was there.
Setting it to `95%` seems to have given it a more accurate margin for the commit view and the PR view.